server/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
numpy==2.1.3
pillow==11.1.0
scikit-image==0.25.2
